requirements.txt
alembic==1.10.2
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6
click==8.1.3
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-CLI==0.4.0
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
pip-autoremove==0.10.0
psycopg2==2.9.6
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.8
typing_extensions==4.5.0
Werkzeug==2.2.3
